Wip search for patient.

+++ actions/deletePatientAction.php
@@ -1,10 +1,13 @@
 <?php
-
+if ($_SERVER['REQUEST_METHOD'] != "POST") {
+    header("Location:../views/dashboard.php");
+    die();
+}
 require_once __DIR__."/../classes/DB.php";
 require_once __DIR__."/../classes/Patient.php";
 use clinic\Patient;
 
-$selectedPatient = Patient::find($_POST['patient_id'],'object');
+$selectedPatient = Patient::findById($_POST['patient_id'],'object');
 $res = $selectedPatient->delete();
 
 if($res) {
